+++ src/Hacks/AutoPractice.cpp
@@ -1,0 +1,19 @@
+#include <Geode/Geode.hpp>
+#include <Geode/modify/PlayLayer.hpp>
+#include "../Client/Client.h"
+
+using namespace geode::prelude;
+
+class $modify (PlayLayer)
+{
+    bool init(GJGameLevel* level, bool useReplay, bool dontCreateObjects)
+    {
+        if (!PlayLayer::init(level, useReplay, dontCreateObjects))
+            return false;
+
+        if (Client::GetModuleEnabled("auto-practice"))
+            this->togglePracticeMode(true);
+
+        return true;
+    }
+};